Don't support mass actions
The mass actions can be replaced by using single actions multiple times.
Though mass actions are useful to reduce API calls and this is critical
in some situations, such situations seem not to be so much. In these
cases, it seems to be better and more beneficial to keep implementation
simple.