Add track/errors.go and track/internal/errors.go
The error definitions in track/toggl/errors.go are divided into these
files. The interface for users to handle errors is defined in
track/errors.go and exported by the package "track". On the other hand,
the error definitions which are hidden from users are defined in
track/internal/errors.go.